Improve selectinline
