+++ clients/admin-ui/cypress/support/ant-support.ts
@@ -138,18 +138,34 @@ Cypress.Commands.add("applyTableFilter", (columnTitle, filterOption) => {
     .siblings(".ant-dropdown-trigger")
     .click();
 
-  // Wait for the filter dropdown to appear
-  cy.get(".ant-table-filter-dropdown").should("be.visible");
+  // Wait for the filter dropdown to appear and find the visible one
+  cy.get(".ant-table-filter-dropdown:visible")
+    .should("be.visible")
+    .within(() => {
+      // Wait for menu items to be available
+      cy.get(".ant-dropdown-menu-item").should("have.length.at.least", 1);
 
-  // Select the filter option
-  if (typeof filterOption === "string") {
-    cy.get(".ant-dropdown-menu-item").contains(filterOption).click();
-  } else {
-    cy.get(".ant-dropdown-menu-item").eq(filterOption).click();
-  }
+      // Select the filter option
+      if (typeof filterOption === "string") {
+        cy.get(".ant-dropdown-menu-item")
+          .contains(filterOption)
+          .should("exist")
+          .click();
+      } else {
+        cy.get(".ant-dropdown-menu-item")
+          .eq(filterOption)
+          .should("exist")
+          .click();
+      }
 
-  // Click OK to apply the filter
-  cy.get(".ant-btn-primary").contains("OK").click();
+      // Click OK to apply the filter
+      cy.get(".ant-table-filter-dropdown-btns .ant-btn-primary")
+        .should("exist")
+        .click();
+    });
+
+  // Wait for the dropdown to disappear
+  cy.get(".ant-table-filter-dropdown:visible").should("not.exist");
 });
 
 export {};
