--- clients/admin-ui/src/features/privacy-requests/ManualTaskTab.tsx
@@ -1,7 +1,0 @@
-import { ManualTasks } from "~/features/manual-tasks/ManualTasks";
-
-const ManualTaskTab = () => {
-  return <ManualTasks />;
-};
-
-export default ManualTaskTab;
